update prompt tests
